+++ pkg/monitoring/exporter/reportobservations_test.go
@@ -3,10 +3,8 @@ package exporter
 import (
 	"testing"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
-	"go.uber.org/zap/zapcore"
 
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 	commonMonitoring "github.com/smartcontractkit/chainlink-common/pkg/monitoring"
@@ -16,35 +14,23 @@ import (
 	"github.com/smartcontractkit/chainlink-solana/pkg/monitoring/types"
 )
 
-func TestReportObservations(t *testing.T) {
+func TestSlotHeight(t *testing.T) {
 	ctx := utils.Context(t)
-	lgr, logs := logger.TestObserved(t, zapcore.ErrorLevel)
-	m := mocks.NewReportObservations(t)
-	m.On("SetCount", mock.Anything, mock.Anything).Once()
-	m.On("Cleanup", mock.Anything).Once()
+	m := mocks.NewSlotHeight(t)
+	m.On("Set", mock.Anything, mock.Anything, mock.Anything).Once()
+	m.On("Cleanup").Once()
 
-	factory := NewReportObservationsFactory(lgr, m)
+	factory := NewSlotHeightFactory(logger.Test(t), m)
 
 	chainConfig := testutils.GenerateChainConfig()
-	feedConfig := testutils.GenerateFeedConfig()
-	exporter, err := factory.NewExporter(commonMonitoring.ExporterParams{ChainConfig: chainConfig, FeedConfig: feedConfig, Nodes: []commonMonitoring.NodeConfig{}})
+	exporter, err := factory.NewExporter(commonMonitoring.ExporterParams{ChainConfig: chainConfig})
 	require.NoError(t, err)
 
 	// happy path
-	exporter.Export(ctx, []types.TxDetails{{ObservationCount: 10}})
+	exporter.Export(ctx, types.SlotHeight(10))
 	exporter.Cleanup(ctx)
 
-	// not txdetails type - no calls to mock
-	assert.NotPanics(t, func() { exporter.Export(ctx, 1) })
-
-	// zero txdetails - no calls to mock
-	exporter.Export(ctx, []types.TxDetails{})
-
-	// empty txdetails
-	exporter.Export(ctx, []types.TxDetails{{}})
-	assert.Equal(t, 1, logs.FilterMessage("exporter could not find non-empty TxDetails").Len())
-
-	// multiple TxDetails should only call for the first non-empty one
-	m.On("SetCount", uint8(1), mock.Anything).Once()
-	exporter.Export(ctx, []types.TxDetails{{}, {ObservationCount: 1}, {ObservationCount: 10}})
+	// test passing uint64 instead of SlotHeight - should not call mock
+	// SlotHeight alias of uint64
+	exporter.Export(ctx, uint64(10))
 }
